Corrects backup script permissions problem introduced by commit 53d9cde
docker_backup.sh contains embedded "sudo" commands so the intention
seems to be that the script should not be run as root. The unprivileged
"touch" command to create the log file in the backups directory fails
because the directory has root ownership. The backup files (.tar.gz) are
given user ownership so the log file should probably have matching user
ownership. This PR reinstates "sudo touch $logfile" but then changes
$logfile ownership to pi:pi before the backup file name is echoed to
the log file.