Update menu.sh
Version "2" of the docker-compose.yml cannot use "Healthcheck" instructions.  Only 2.1+ can do this.  I propose we upgrade the version to at least 3 so we can incorporate HealthCheck commands and any other new features as appropriate.  We're running the latest version of docker - the latest version of the docker-compose.yml is 3.7 - unless there's a really, really good reason to use an older format of the docker-compose.yml, can we please update it to at least 3.0, preferably 3.7?