Fixes some inconsistencies and makes other tweaks to index page
Removed references to the Wiki (this IS the Wiki).

Fixed typo in EspruinoHub (was EspurinoHub).

Changed "sudo apt-get install git" to "sudo apt install git". My
understanding is that "apt" is considered best practice.

Changed the URL to download the repository from
"https://github.com/gcgarner/IOTstack.git" to
"https://github.com/SensorsIot/IOTstack.git".

Removed the last sentence under "Running under a virtual machine".
Graham closed and unpinned issue 29 so this seems redundant.

Changed several URLs to refer to the new documentation structure.

Added a "Networking" section below "Persistent data" to cover-off the
"first timer" problem of not realising why 127.0.0.1 won't work. It's
such a common problem that I thought it would be better to put it up
front (and thereby explain it twice) rather than expect newbies to
read the Networking page and immediately "grok" its significance.

Other tweaks such as clarifying the need to install Docker on the
first menu run, and consistent layout for the commands to start and
stop the stack. Also harmonised "IOTstack" as the spelling and added
some missing full stops.