Mosquitto - make security persistent + extend documentation
Implements changes proposed in issue #68:

1. Add a volume mapping to the template `service.yml`:

	```
	- ./volumes/mosquitto/pwfile:/mosquitto/pwfile
	```

2. Add this line to the template `directoryfix.sh`:

	```
	sudo mkdir -p ./volumes/mosquitto/pwfile/
	```

3. Change the relevant line in the template `mosquitto.conf` to read:

	```
	#password_file /mosquitto/pwfile/pwfile
	```

4. Change the documentation so that the command reads:

	```
	mosquitto_passwd -c /mosquitto/pwfile/pwfile «MYUSER»
	```

Extends Mosquitto documentation to:

* explain logging options
* elaborate on setting up security
* explain how to run Mosquitto as root (rather than user 1883) in
situations where that might be appropriate

Also adds a timestamp format line to the template `mosquitto.conf`:

```
log_timestamp_format %Y-%m-%dT%H:%M:%S
```

This causes log entries to have human-readable timestamps. The
timestamps are UTC which is, I think, a huge improvement on seconds
since 1/1/70. It would be nicer if the timestamps could be converted
to local time but I could not find a configuration option to make
that happen (setting TZ didn't do the trick either).