Update to Blynk local server including config files - 877dev
The original version did not create any config files and did not support email.

This commit fixes that, by making use of `directoryfix.sh` to create the config folder,
which contains two config files, `server.properties` and `mail.properties`.
Note that if they already exits they are not overwritten.

The Dockerfile has been updated to use the latest version 0.41.14,
and a parameter to use a custom mail.properties.

Server.properties should work "as is" for most people, but can be changed as required.
Mail.properties requires editing to add gmail credentials.
Service.yml updated to map the configs to the volumes folder.

A new readmeguide has been produced which needs adding to the main webpage, this can be found at
https://gist.github.com/877dev/cfa52044336e1e44d74156ce8b1458aa#file-blynk-setup-guide-md