Adds template and documentation for Portainer-ce
This PR proposes supporting Portainer and Portainer CE 2.0 in parallel.
See also [Issue 132](https://github.com/SensorsIot/IOTstack/issues/132).

According to [Portainer documentation]
(https://www.portainer.io/2020/08/portainer-ce-2-0-what-to-expect/):

> Portainer 1.24.x will continue as a separate code branch, released as
> portainer/portainer:latest, and will receive ongoing security updates
> until at least 1st Sept 2021. No new features will be added beyond
> what was available in 1.24.1

The template for Portainer CE:

- Maps external port 9002 to internal port 9000.
This is so Portainer CE can coexist with Portainer.
- Maps external port 8000 to internal port 8000.
Portainer did not expose port 8000.

[Portainer documentation]
(https://www.portainer.io/2019/07/how-does-the-edge-agent-work/)
defines Port 8000:

> Port 8000 is a SSH tunnel server and is used to create a secure tunnel
> between the agent and the Portainer instance.

The template for Portainer CE includes `directoryfix.sh` which copies:

```
~/IOTstack/volumes/portainer
```

to:

```
~/IOTstack/volumes/portainer-ce
```

if and only if the former exists **and** the latter does not exist. This
is intended as a one-time assist for users migrating between Portainer
and Portainer-CE.

Also includes documentation page for Portainer-CE.