Try to head off problems created by unnecessary use of sudo
A recurrent theme on Discord is questions where the underlying cause
of the user's problem is an unnecessary use of the sudo command.

To try to reduce the incidence of both the problems and the questions,
this PR adds code to the start of most bash scripts. The code checks
the EUID and exits with "This script should NOT be run using sudo".

A discussion of the problem of using sudo is also added to
docs/Getting-Started.md. Aside from hopefully catching the eye of
every new IOTstack user, this will be a useful anchor for Discord
questions.

Some scripts also lacked the "hash bang" preamble and I took the
opportunity to fix such problems.