2020-12-10-docker_backup.sh folder fix
Dropbox are about to change API versions and discontinue the current API's in 2021.
The developer of Dropbox-Uploader has updated the project, it seems for this reason.
The default way the files are saved has been changed, hence this PR is need to fix that.
You will need to create a new dropbox_uploader.sh config file if not sone already ready for
the discontinuation of the old API. https://github.com/andreafabrizi/Dropbox-Uploader

Also updated Dropbox file retention to 30 which is more useful for restoring after an un-noticed issue.