Adds service definition for AdGuard Home - master (1 of 2)
Adds service definition and documentation for AdGuard Home.

I did look, briefly, at the question of a "build.py" file but decided
that that question should be left for someone who understands if or
when it is necessary, and what is needed in it if it is. That someone
is not me. There are several templates that lack a "build.py"
(dashmachine, domoticz, heimdall, homer, portainer_agent, qbittorrent)
so it does not seem to be essential. It shows up in new menu OK. There
are no options that need to be configured in advance.

I made a conscious choice with port 8089 for the admin interface port
**because** it conflicts with PiHole. There are already going to be
other conflicts (eg 53) so this just rams home the point that you
can't run both.

Conversely, a search suggested no conflict with port 3001.