Adds web UI password field to Dockerfile - master (1 of 2)
Adds a commented-out "auth_token" field to the default configuration
file set up by the Dockerfile. This follows on from discussion at
[PR310](https://github.com/SensorsIot/IOTstack/pull/310#issuecomment-816557323).

The basic idea is that the web UI should work out-of-the-box but it
should also be easy for anyone who wants to add basic security to
the web interface to do that (or disable the web UI entirely).

Documentation updated to reflect these changes.