Improved mkdocs.yml - deactivate awesome-pages
`awesome-pages` was in the wrong place in `mkdocs.yml` but was not
producing errors. Moving it to the correct place in PRs 335 and 336
revealed that `mkdocs-awesome-pages-plugin` had not been installed so
documentation failed to build.

This patch comments-out `awesome-pages` on the basis that is seems to
make no real difference to the displayed pages.