20210924 Portainer - master branch - PR 1 of 3
In [Issue 289](https://github.com/SensorsIot/IOTstack/issues/289),
@wijnsema proposed removal of Portainer. There has been additional
support for this from @mats-nk. It is coming up to a year since
Portainer-CE was added to IOTstack so this PR is my way of saying that
I agree with the idea.

1. Remove Portainer service definition.
2. Remove Portainer documentation.
3. Alter Portainer-CE documentation to remove references to Portainer.

The `build.py` could probably be removed too but it looked to me like
it would function as a no-op so I left it alone.