20211014 NextCloud documentation
The `mkdocs` rendering of the NextCloud container documentation page was
not starting a new numbered list item for step 10.

I traced the problem to extraneous tab characters.

Renderings by GitHub (inline) and other engines (MacDown, QLMarkdown)
were correct so this is another example of either `mkdocs` or the
`material` theme being hypersensitive.