docs: console-highlight: fix copy button
Normally the copy-button at the top-right of every code-block will copy
all of its contents. This isn't wanted for code-blocks of the syntax
'console'. In these blocks only want to copy-to-clipboard the actual
commands, but skip prompts and command outputs.

The previous commit makes such elements non-selectable. This builds on
that to only copy the user-selectable parts of the code-block. What to
copy is overridden by setting the 'data-clipboard-text' attribute.