20220708 Node-RED - master branch - PR 1 of 2 - part 2
Fixes `nodered_list_installed_nodes.sh`. Now relies on `npm` running
inside the container to discover what's installed where.

Documentation describing use of `/data` on `npm` calls turns out to be
ineffective. The solution is to set the working directory on the
`docker exec` call.

Following on from a discussion on Discord
[starting here](https://discord.com/channels/638610460567928832/638610461109256194/994818372439113838),
it was obviously confusing for the documentation to be discussing
"nodes" when `npm` can be used to install a variety of components.
Documentation adjusted to clarify this.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>