20220811 ZeroTier - master branch - PR 1 of 2
Adds first cut of ZeroTier in the form of two containers:

* ZeroTier-client from `zyclonite:zerotier`
* ZeroTier-router from `zyclonite:zerotier-router`

The two containers are mutually exclusive (no menu check - docker-compose
will complain about port conflicts). My guess is that IOTstack users will
generally prefer the router because it is closer to WireGuard
functionality, and is likely what an IOTstack user will be looking for
if they find themselves stuck behind CGNAT.

ZeroTier-client is the standard client. Has no environment variables or
other preconfiguration. Needs to be instructed to join the user's
ZeroTier network after first launch. Otherwise works "as is".

ZeroTier-router is built on top of ZeroTier-client and adds management
of iptables rules at startup/shutdown to support routing to/from the
local network. Has environment variables with defaults appropriate for
IOTstack.

Adds documentation:

* ZeroTier.md
* ZeroTier-vs-WireGuard.md to explain the pros/cons of each.

Modifies WireGuard documentation to reference ZeroTier-vs-WireGuard.md
and also adds a tweak to one diagram (takes the opportunity to switch
from PNG to JPEG ; PNGs do not always "play nice" when Wiki pages are
viewed in Dark Mode).

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>