20220913 Influx documentation - master branch
Fixes the problem reported by #599.

This is an unfortunate, predictable and predicted side-effect of moving
from this style of anchored title:

```
```

to this style:

```
```

The former style is testable by generating HTML from the Markdown and
running it through a validator. The latter style isn't amenable to that
approach because anchor generation is done "just in time" by mkdocs.
Until we find some way to test mkdocs output in a systematic way,
fairly trivial semantic errors (like the missing "#" in this case) will
occasionally slip through the cracks.

Also fixed another broken link (`#authWarning`).

Fixes #599.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>