2022-09-25 Zigbee2MQTT - master branch
Clarify what happens if the container starts with the default device
`/dev/ttyAMA0` in the service definition. The process probes the device,
finds it doesn't respond like a Zigbee adapter, and aborts without
starting the web GUI. Because of the `restart: unless-stopped` clause,
Docker restarts the container and the same sequence repeats.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>