document host.docker.internal
Documents host.docker.internal as an alternative to 172.17.0.1, which
is an alternative to an IP address of a physical network interface or
fully-qualified domain name for an interface.

Uses host.docker.internal in the GPIO and SSH examples.

SSH example simplified. Now only describes calling out of the Node-RED
container to the host where the container is running.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>