2023-05-16 Home Assistant adaptation to expected kernel change - master branch - PR 1 of 2
Background:

- #690 – Kernel update may remove /dev/ttyAMA0
- [Home Assistant documentation](https://www.home-assistant.io/installation/raspberrypi#docker-compose)

Changes:

1. Removes Bluetooth-specific volumes and devices from template.
2. Removes Pi-hardware-specific image options (still available but no
longer mentioned in the Home Assistant documentation; and not relevant
where IOTstack may be deployed on non-Pi hardware).
3. Adds documentation to Home Assistant wiki page (master branch) to
explain how to re-add volumes and devices if Bluetooth support is
required on a Raspberry Pi:

	- This means Home Assistant will launch successfully on non-Pi
hardware.
	- Documentation references `/dev/serial1` rather than `/dev/ttyAMA0`
as the external device but maintains the latter as the internal device.

3. Removes documentation from the Home Assistant wiki page (master
branch) where Pi-hardware-specific image options were discussed.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>