2023-06-15 Pi-hole - master branch - PR 1 of 2
Applies changes recommended by `yamllint`.

Adds `FTLCONF_MAXDBDAYS` to the template with the existing default of
365 days, and adds words to the documentation explaining its use.

Adds `PIHOLE_DNS_` with the existing defaults of 8.8.8.8 and 8.8.4.4,
and adds words and screenshot to the documentation explaining its use.

Explains issue with Docker Desktop for macOS raised in #709 where all
client activity is logged against the default gateway on the internal
bridged network.

Tidies-up problematic code-fence language specs. Shortens some
subheadings so they fit on a single line in the Wiki TOC.

Fixes #709

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>