2023-07-03 Python requirements files + Bookworm - master branch
The underlying cause of the problem reported in #660 is the Python
requirements files containing pins for (almost) every line item.

I appreciate that pinning is what happens by default when requirements
files are generated but some of the pinned versions which might have
been appropriate for Bullseye when the files were generated will not
actually install on Bookworm.

This Pull Request removes pins from every line item in each requirements
file. In other words, the defaults will become the latest and greatest
for each system (Buster or Bullseye or Bookworm).

I have tested this on Bullseye and Bookworm (but not Buster), and only
to the extent of confirming that each listed package gets installed
and the menu doesn't chuck up any complaints. The test systems were
built with [PiBuilder](https://github.com/Paraphraser/PiBuilder).

**IF** we encounter a problem which needs to be solved by pinning a
package to a specific version, we will have to sort out how to do that
across OS releases.

Fixes: #660

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>