2023-10-09 Mosquitto entry-point script - master branch - PR 1 of 2
The folder structure for a newly-created or self-repaired persistent
store assumes ownership of id 1883 (which is "mosquitto" inside the
container) and file/folder modes copied from the `iotstack_defaults`
folder hierarchy.

Investigations during #731 revealed that Mosquitto now wants the
password file to be owned by root with a mode of 600.

Although the mode requirement could be accommodated by changing the
`iotstack_defaults` structure, the ownership requirement can't be met
that way.

This fix adds the necessary `chown` and `chmod` commands to the
entry-point script.

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>