2023-10-14 documentation - master branch
Replaces all example commands in the following format:

```
$ docker-compose rm --force --stop -v «container»
```

with example commands in the this format:

```
$ docker-compose down «container»
```

The Getting Started document includes a note explaining that any error
returned by `docker-compose down` to the effect that it does not support
`«container»` arguments implies that the user's version of
`docker-compose` is obsolete and should be updated.

Where any container's documentation has had `rm` replaced with `down`,
that is followed by a note pointing to the Getting Started explanation.

Also, I took the opportunity to fix the following error returned by
`mkdocs`:

```
INFO    -  Doc file 'index.md' contains an unrecognized relative link 'Developers/', it was left as is. Did you mean 'Developers/index.md'?
```

Signed-off-by: Phill Kelley <34226495+Paraphraser@users.noreply.github.com>